requirements.txt
# --- Core Backend ---
fastapi==0.128.1
uvicorn==0.40.0
python-multipart==0.0.22
pydantic==2.12.5
python-dotenv==1.2.1

# --- RAG & AI Engine ---
langchain==0.3.14
langchain-community==0.3.14
langchain-huggingface==0.2.0
langchain-qdrant==0.2.1
langchain-groq==0.3.2
qdrant-client==1.16.2
flashrank==0.2.10
sentence-transformers==5.2.2
groq==0.37.1

# --- Document Processing ---
unstructured==0.18.31
python-docx==1.2.0
pypdf==6.6.2
pdf2image==1.17.0
pdfminer.six==20260107

# --- Frontend ---
streamlit==1.54.0
requests==2.32.5